docs: Add preliminary risk analysis
This adds a paragraph on the risks associated with the system.
Since biases are not strictly illegal, and even expected in any real-world income data,
we instead mention that downstream users have to take care when applying the system to
their own uses, in order to not perpetuate the detected biases unwittingly.