deploy: Add docker compose file for lakeFS + minIO S3
Creates a local lakeFS server with a minIO S3 deployment.
In the minIO setup, a bucket named "lakefs" is created on the server, which
is also given as the default storage namespace prefix in the server options,
so that the user does not need to specify their own prefix in repo creation.