deploy: Add docker compose file for lakeFS + minIO S3 (#27)
* deploy: Add docker compose file for lakeFS + minIO S3

Creates a local lakeFS server with a minIO S3 deployment.
In the minIO setup, a bucket named "lakefs" is created on the server, which
is also given as the default storage namespace prefix in the server options,
so that the user does not need to specify their own prefix in repo creation.

* deploy: Move lakeFS service into local compose stack

* chore: Rename Docker Compose stack

* feat: Introduce LakeFS I/O manager

The manager stores Pandas DataFrames a Parquet files in
a lakeFS repository, based on the Dagster UPath I/O mgr.

* deploy: Modularize Docker Compose stack through profiles

* deps: Re-lock dependencies

* fix: Simplify path handling in lakeFS I/O manager

---------

Co-authored-by: Adrian Rumpold <a.rumpold@gmail.com>