feat: Implement Dagster sensors for model evaluation & data quality reports (POC)
The model evaluation sensor checks the MLflow model registry
for a new model version and starts an asset job that performs
the performance evaluation based on the test dataset.

The Dagster assets have been reorganized into groups for
easier visual identification.

The data quality report sensor is a very basic POC to show how
the loop between the FastAPI application and the inference
tables needed for NannyML report generation in Dagster could
be closed. There be dragons.