feat: Record canonical URI in lakeFS I/O manager
When committing data in the I/O manager, the resulting commit hash
is recorded in the output metadata, along with the immutable ref
to the written data.

This allows for easier reconstruction of the data at the time of
the pipeline execution, since the ref containing the branch name
might have changed in the meantime.