docs: Fix link definitions
In Markdown, link definitions (where the URL is defined separately from
the link text) must be followed by a second square bracket containing
the ID (or empty to use the link text as the ID).

In some places, the Markdown parser seems to accept the links without
the IDs, but in the index table it leads to broken links and missing
items in the table (because the following link definition is treated as
the ID for the previous link).

See this page for more details:
https://daringfireball.net/projects/markdown/syntax#link