docs: Add small component synopsis to system overview page (#88)
* docs: Add small component synopsis to system overview page

The purpose of these small synopses is that the interested reader is being presented
with some information on the basic functionalities and technology choices of our system,
without immediately having to click away into the engineering practices, which is not
good for maintaining attention span and engagement with the content.

* docs: Add tool links to system overview

Also update a reference in the model serving page to link to
the mlserver docs only, which by default redirects to the stable
release documentation.