feat: Refactor Dagster pipeline with explicit dependencies for model/monitoring container assets
This allows the pipeline to be executed as a single job,
whereas before the container images were only built
after the sensor picked up a new model version in the
model registry.

This simplifies the execution of the pipeline at the expense of
having to modify the upstream dependencies of the container
assets if a new model architecture is added as a separate
asset.