Add training pipeline runner script (#104)
* feat: Add Python script for e2e training trigger

* deps: Update containerized MLflow server to 2.22.0

* fix: Don't allow prerelease dependencies in model serving container image

* feat: Refactor Dagster pipeline with explicit dependencies for model/monitoring container assets

This allows the pipeline to be executed as a single job,
whereas before the container images were only built
after the sensor picked up a new model version in the
model registry.

This simplifies the execution of the pipeline at the expense of
having to modify the upstream dependencies of the container
assets if a new model architecture is added as a separate
asset.

* fix: Install git in docker image to pull nannyML dependency from git

* deps: Align Dagster dependency versions in containers with pyproject.toml

* fix(scripts): Invoke Dagster training pipeline through GraphQL API

---------

Co-authored-by: Kristof Schroeder <kristof_schroeder@web.de>