docs: Remove two TODOs in article mapping of engineering practices
While they have not been expanded a lot, both of them now contain another
subpoint to give the reader a direction of where to look for additional information.

In the case of model serving, the current list is at one subparagraph of Art. 15,
but more can be added in the future.