fix: Use fastparquet in lakeFS I/O manager
The `pyarrow` engine does not preserve categorical dtypes with numeric
data, which can lead to issues when reading and writing Parquet files.

The `fastparquet` engine is now used for reading and writing Parquet
files in the LakeFS I/O manager, to ensure a consistent representation
of data types, especially for categorical features.